Remove g++ warnings.

+++ src/Packages/Uintah/Core/Datatypes/NCTensorField.cc
@@ -30,7 +30,7 @@ NCTensorField::NCTensorField(const NCTensorField& copy)
    high(copy.high), low(copy.low), nx(copy.nx),
    ny(copy.ny), nz(copy.nz)
 {
-  for(int i = 0; i < copy._vars.size(); i++){
+  for(int i = 0; i < (int)copy._vars.size(); i++){
     _vars.push_back( copy._vars[i] );
   }
  
@@ -46,7 +46,7 @@ NCTensorField::NCTensorField(GridP grid, LevelP level,
    high(-MAXINT,-MAXINT,-MAXINT),
    low(MAXINT,MAXINT,MAXINT)
 {
-  for(int i = 0; i < vars.size(); i++){
+  for(int i = 0; i < (int)vars.size(); i++){
     _vars.push_back( vars[i]);
   }
   computeHighLowIndices();
